biocomputing/harness/pyproject.toml
[project]
name = "harness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "libroadrunner>=2.7.0",
    "matplotlib>=3.10.0",
    "plotly>=6.0.0",
]
